Removed TrackingService class, added interval location information into main fragment
